modify: move commands to correspond class
